enable all tests
